refactor (front): renomeando Profile -> ProfileModal e colocando Icons em assets
