+++ front/src/renderer/src/assets/Icons/Business.tsx
@@ -1,0 +1,6 @@
+
+export function BusinessIcon() {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" width="32" height="32" fill="none"><rect width="30.5" height="30.5" x=".75" y=".75" fill="#E7F5FF" rx="15.25"/><rect width="30.5" height="30.5" x=".75" y=".75" stroke="#E7F5FF" strokeWidth="1.5" rx="15.25"/><path stroke="#1864AB" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.09091" d="m9.97961 13.4925 5.72349-3.69497c.1679-.1083.3845-.1083.5523 0l5.7242 3.69497c.1449.0941.2328.2552.2328.4284v.4014c0 .2809-.2281.509-.5096.509h-11.447c-.2816 0-.50971-.2281-.50971-.509v-.4014c0-.1732.088-.3343.23352-.4284ZM15.9785 12.6946v.0066"/><path stroke="#1864AB" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.09091" d="m22.2311 21.85-.27-1.2258c-.0359-.1631-.1808-.2795-.3479-.2795h-11.269c-.1672 0-.3121.1164-.34795.2795L9.72618 21.85c-.04873.222.12048.4325.34792.4325h11.8098c.2274 0 .396-.2105.3472-.4325Z" clipRule="evenodd"/><path stroke="#1864AB" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.09091" d="M17.5536 14.8301v5.5151m3.1516-5.5151v5.5151m-9.4562-5.5151v5.5151m3.1516-5.5151v5.5151"/></svg>
+  );
+}
