+++ front/src/renderer/src/assets/Icons/ClientIcon.tsx
@@ -1,0 +1,6 @@
+
+export function ClientIcon() {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" width="32"  height="32"  fill="none"><rect width="30.5" height="30.5" x=".75" y=".75" fill="#E7D6F1" rx="15.25"/><rect width="30.5" height="30.5" x=".75" y=".75" stroke="#E7D6F1" strokeWidth="1.5" rx="15.25"/><path fill="#53405E" d="M15.9997 10.5625c-.2216 0-.441.0436-.6458.1285-.2047.0848-.3907.2091-.5474.3658s-.281.3427-.3658.5474c-.0848.2048-.1285.4242-.1285.6458 0 .2216.0437.441.1285.6458.0848.2047.2091.3907.3658.5474s.3427.281.5474.3658c.2048.0849.4242.1285.6458.1285.4476 0 .8768-.1778 1.1933-.4943.3164-.3164.4942-.7456.4942-1.1932s-.1778-.8768-.4942-1.1932c-.3165-.3165-.7457-.4943-1.1933-.4943ZM13.1872 12.25c0-.7459.2963-1.4613.8238-1.9887.5274-.52748 1.2428-.8238 1.9887-.8238s1.4613.29632 1.9888.8238c.5274.5274.8237 1.2428.8237 1.9887s-.2963 1.4613-.8237 1.9887c-.5275.5275-1.2429.8238-1.9888.8238s-1.4613-.2963-1.9887-.8238c-.5275-.5274-.8238-1.2428-.8238-1.9887Zm.798 4.3643c.099.0735.2265.1597.3788.2452.3847.2175.9487.4425 1.635.4425.6862 0 1.251-.225 1.6357-.4425.1523-.0855.2798-.1717.3788-.2452.141.0375.2807.0789.4192.1244l.72.2363c.54.1778.948.609 1.0935 1.1423l.2978 2.148c.0562.408-.1575.7042-.4545.7747-.7853.1875-2.109.3975-4.0898.3975-1.9807 0-3.3052-.21-4.0905-.3975-.2962-.0705-.51-.3667-.4545-.7747l.2978-2.148c.0739-.266.2126-.5094.4035-.7087.1909-.1993.4282-.3483.6907-.4336l.72-.2363c.138-.045.2775-.0864.4185-.1244Zm.5325-1.0163-.2182-.2152-.3015.0689c-.263.0601-.524.1328-.783.2183l-.72.2362c-.4473.1454-.8506.4016-1.1724.7446-.3218.343-.5516.7618-.6681 1.2174l-.0075.0301-.306 2.2125c-.1208.873.3465 1.7947 1.308 2.0235.8775.2092 2.289.4282 4.35.4282 2.0617 0 3.4732-.219 4.3507-.4275.9615-.2295 1.4295-1.1512 1.3088-2.0235l-.3068-2.2125-.0075-.03c-.1165-.4559-.3464-.8749-.6683-1.2181-.3219-.3431-.7254-.5993-1.1729-.7447l-.7193-.2362c-.2585-.085-.5195-.1575-.783-.2175l-.3015-.0697-.2175.2144-.0007.0008c-.029.0264-.0591.0517-.09.0757-.0967.0769-.1996.1456-.3075.2055-.3298.1906-.703.2932-1.0838.2978-.3808-.0046-.754-.1072-1.0837-.2978-.1417-.08-.2757-.1742-.3983-.2812Z"/></svg>
+  );
+}
