+++ front/src/renderer/src/assets/Icons/CloseIcon.tsx
