+++ front/src/renderer/src/assets/Icons/DeleteRoundedIcon.tsx
@@ -1,0 +1,5 @@
+export function DeleteRoundedIcon() {
+  return (
+    <svg xmlns="http://www.w3.org/2000/svg" width="52" height="52" fill="none"><rect width="52" height="52" fill="#FFF5F5" rx="26"/><path stroke="#C92A2A" strokeLinecap="round" strokeLinejoin="round" strokeWidth="1.5" d="M33.3247 23.4675s-.543 6.735-.858 9.572c-.15 1.355-.987 2.149-2.358 2.174-2.609.047-5.221.05-7.829-.005-1.319-.027-2.142-.831-2.289-2.162-.317-2.862-.857-9.579-.857-9.579M34.708 20.239H17.75m13.6904 0c-.785 0-1.461-.555-1.615-1.324l-.243-1.216c-.15-.561-.658-.949-1.237-.949h-4.233c-.579 0-1.087.388-1.237.949l-.243 1.216c-.154.769-.83 1.324-1.615 1.324"/></svg>
+  );
+}
