--- front/src/renderer/src/views/Icons/ClientIcon/index.tsx
@@ -1,20 +1,0 @@
-import { Container } from './styles';
-
-export function ClientIcon() {
-  return (
-    <Container>
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="18"
-        height="18"
-        viewBox="0 0 18 18"
-        fill="none"
-      >
-        <path
-          d="M8.99972 3.5625C8.77811 3.5625 8.55867 3.60615 8.35394 3.69095C8.1492 3.77576 7.96317 3.90006 7.80647 4.05676C7.64977 4.21346 7.52547 4.39948 7.44067 4.60422C7.35586 4.80896 7.31222 5.02839 7.31222 5.25C7.31222 5.47161 7.35586 5.69104 7.44067 5.89578C7.52547 6.10052 7.64977 6.28654 7.80647 6.44324C7.96317 6.59994 8.1492 6.72424 8.35394 6.80905C8.55867 6.89385 8.77811 6.9375 8.99972 6.9375C9.44727 6.9375 9.87649 6.75971 10.193 6.44324C10.5094 6.12678 10.6872 5.69755 10.6872 5.25C10.6872 4.80245 10.5094 4.37323 10.193 4.05676C9.87649 3.74029 9.44727 3.5625 8.99972 3.5625ZM6.18722 5.25C6.18722 4.50408 6.48353 3.78871 7.01098 3.26126C7.53842 2.73382 8.25379 2.4375 8.99972 2.4375C9.74564 2.4375 10.461 2.73382 10.9885 3.26126C11.5159 3.78871 11.8122 4.50408 11.8122 5.25C11.8122 5.99592 11.5159 6.71129 10.9885 7.23874C10.461 7.76618 9.74564 8.0625 8.99972 8.0625C8.25379 8.0625 7.53842 7.76618 7.01098 7.23874C6.48353 6.71129 6.18722 5.99592 6.18722 5.25ZM6.98522 9.61425C7.08422 9.68775 7.21172 9.774 7.36397 9.8595C7.74872 10.077 8.31272 10.302 8.99897 10.302C9.68522 10.302 10.25 10.077 10.6347 9.8595C10.787 9.774 10.9145 9.68775 11.0135 9.61425C11.1545 9.65175 11.2942 9.69325 11.4327 9.73875L12.1527 9.975C12.6927 10.1528 13.1007 10.584 13.2462 11.1173L13.544 13.2653C13.6002 13.6733 13.3865 13.9695 13.0895 14.04C12.3042 14.2275 10.9805 14.4375 8.99972 14.4375C7.01897 14.4375 5.69447 14.2275 4.90922 14.04C4.61297 13.9695 4.39922 13.6733 4.45472 13.2653L4.75247 11.1173C4.82645 10.8513 4.96509 10.6079 5.15602 10.4086C5.34695 10.2093 5.58423 10.0603 5.84672 9.975L6.56672 9.73875C6.70472 9.69375 6.84422 9.65225 6.98522 9.61425ZM7.51772 8.598L7.29947 8.38275L6.99797 8.45175C6.73497 8.51175 6.47397 8.5845 6.21497 8.67L5.49497 8.90625C5.04768 9.05164 4.64441 9.30775 4.32263 9.65077C4.00085 9.99379 3.77101 10.4126 3.65447 10.8682L3.64697 10.8983L3.34097 13.1108C3.22022 13.9838 3.68747 14.9055 4.64897 15.1343C5.52647 15.3435 6.93797 15.5625 8.99897 15.5625C11.0607 15.5625 12.4722 15.3435 13.3497 15.135C14.3112 14.9055 14.7792 13.9838 14.6585 13.1115L14.3517 10.899L14.3442 10.869C14.2277 10.4131 13.9978 9.99411 13.6759 9.65095C13.354 9.30778 12.9505 9.0516 12.503 8.90625L11.7837 8.67C11.5252 8.585 11.2642 8.5125 11.0007 8.4525L10.6992 8.38275L10.4817 8.59725L10.481 8.598C10.452 8.62442 10.4219 8.64969 10.391 8.67375C10.2943 8.75057 10.1914 8.8193 10.0835 8.87925C9.75374 9.06981 9.38053 9.17235 8.99972 9.177C8.61891 9.17235 8.2457 9.06981 7.91597 8.87925C7.77427 8.79922 7.64034 8.70498 7.51772 8.598Z"
-          fill="#53405E"
-        />
-      </svg>
-    </Container>
-  );
-}
