--- front/src/renderer/src/views/Icons/ClientIcon/styles.ts
@@ -1,16 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  align-items: center;
-  background: #E7D6F1;
-  border-radius: 50%;
-  display: flex;
-  height: 32px;
-  justify-content: center;
-  width: 32px;
-
-  svg {
-    height: 2.4rem;
-    width: 2.4rem;
-  }
-`;
