--- front/src/renderer/src/views/Icons/EnterpriseIcon/index.tsx
@@ -1,46 +1,0 @@
-import { Container } from './styles';
-
-export function EnterpriseIcon() {
-  return (
-    <Container>
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="18"
-        height="18"
-        viewBox="0 0 18 18"
-        fill="none"
-      >
-        <path
-          d="M2.97961 6.4925L8.70314 2.79753C8.87098 2.68923 9.08756 2.68923 9.2554 2.79753L14.9796 6.4925C15.1245 6.58658 15.2124 6.74767 15.2124 6.92094V7.32232C15.2124 7.60322 14.9843 7.83133 14.7028 7.83133L3.25577 7.83133C2.9742 7.83133 2.74609 7.60322 2.74609 7.32232V6.92094C2.74609 6.74767 2.83409 6.58658 2.97961 6.4925Z"
-          stroke="#1864AB"
-          strokeWidth="1.09091"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-        <path
-          d="M8.97852 5.69458V5.70125"
-          stroke="#1864AB"
-          strokeWidth="1.09091"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-        <path
-          fillRule="evenodd"
-          clipRule="evenodd"
-          d="M15.2311 14.8499L14.9611 13.6241C14.9252 13.461 14.7803 13.3446 14.6132 13.3446L3.34415 13.3446C3.17697 13.3446 3.03212 13.461 2.99625 13.6241L2.72618 14.8499C2.67745 15.0719 2.84666 15.2824 3.07409 15.2824L14.8839 15.2824C15.1113 15.2824 15.2799 15.0719 15.2311 14.8499Z"
-          stroke="#1864AB"
-          strokeWidth="1.09091"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-        <path
-          d="M10.5536 7.82996V13.345M13.7052 7.82996V13.345M4.24902 7.82996L4.24902 13.345M7.40064 7.82996L7.40064 13.345"
-          stroke="#1864AB"
-          strokeWidth="1.09091"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-        />
-      </svg>
-    </Container>
-  );
-}
