+++ front/src/renderer/src/views/components/Fab/index.tsx
@@ -2,8 +2,9 @@ import { LuPlus } from 'react-icons/lu';
 
 import { DropdownMenu } from '../DropdownMenu';
 
-import { ClientIcon } from '@renderer/views/Icons/ClientIcon';
-import { EnterpriseIcon } from '@renderer/views/Icons/EnterpriseIcon';
+import { BusinessIcon } from '@renderer/assets/Icons/Business';
+import { ClientIcon } from '@renderer/assets/Icons/ClientIcon';
+
 import { Container, StyledButton, StyledItem } from './styles';
 
 export default function Fab() {
@@ -25,7 +26,7 @@ export default function Fab() {
           </DropdownMenu.Item>
           <DropdownMenu.Item onSelected={() => console.log('funcoinou')} asChild>
             <StyledItem>
-              <EnterpriseIcon />
+              <BusinessIcon />
               Nova empresa
             </StyledItem>
           </DropdownMenu.Item>
