+++ front/src/renderer/src/views/layouts/NavLayout/index.tsx
@@ -3,14 +3,15 @@ import { Outlet } from 'react-router-dom';
 import Navigation from '@renderer/views/components/Navigation';
 
 import { ModalsProvider } from '@renderer/app/contexts/ModalsContext';
-import Profile from '@renderer/views/pages/Profile';
+import ProfileModal from '@renderer/views/modals/ProfileModal';
+
 import { Container } from './styles';
 
 export default function NavLayout() {
   return (
     <ModalsProvider>
       <Container>
-        <Profile />
+        <ProfileModal />
         <Navigation />
         <Outlet />
       </Container>
